964: Renamed Interactive to InteractiveSlide
