+++ src/Controller/InteractiveController.php
@@ -8,7 +8,7 @@ use App\Entity\ScreenUser;
 use App\Entity\Tenant\Slide;
 use App\Entity\User;
 use App\Exceptions\NotFoundException;
-use App\Service\InteractiveService;
+use App\Service\InteractiveSlideService;
 use Symfony\Bundle\SecurityBundle\Security;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
@@ -18,7 +18,7 @@ use Symfony\Component\HttpKernel\Attribute\AsController;
 final readonly class InteractiveController
 {
     public function __construct(
-        private InteractiveService $interactiveSlideService,
+        private InteractiveSlideService $interactiveSlideService,
         private Security $security,
     ) {}
